backend/requirements.txt
aiohttp==3.7.3
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==3.0.1
attrs==23.2.0
black==24.4.2
certifi==2024.2.2
chardet==3.0.4
click==8.1.7
dacite==1.6.0
distro==1.9.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
exceptiongroup==1.2.1
h11==0.14.0
honcho==1.1.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
mypy-extensions==1.0.0
openai==1.30.3
packaging==24.0
pathspec==0.12.1
pcpartpicker==2.2.2
platformdirs==4.2.2
py-moneyed==0.8.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.0
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.12.0
yarl==1.9.4
